requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.29.0

python-dotenv==1.0.1

gspread==6.1.4
oauth2client==4.1.3

pydantic==2.6.1

# FSM storage
redis==5.0.1   # можно не использовать, но пусть будет для будущего апгрейда
